Use the size_t modifier to avoid an unnecessary cast
(cherry picked from commit 8f49a55c03ea8a2340f56124567bbe610906df9d)

+++ modules/db_postgres/pg_con.c
@@ -60,7 +60,7 @@ struct pg_con* db_postgres_new_connection(struct db_id* id)
 				"\n", (unsigned long)sizeof(struct pg_con));
 		return 0;
 	}
-	LM_DBG("%p=pkg_malloc(%lu)\n", ptr, (unsigned long)sizeof(struct pg_con));
+	LM_DBG("%p=pkg_malloc(%zu)\n", ptr, sizeof(struct pg_con));
 
 	memset(ptr, 0, sizeof(struct pg_con));
 	ptr->ref = 1;
